refactor: implement remaining architectural improvements
Decimal for monetary values:
- Domain/application layers: all price fields changed float → Decimal
- Presentation request schemas use Decimal (Pydantic handles JSON → Decimal)
- Response schema (ItemResponse) uses float — explicit Decimal → float
  conversion in ItemSchemaMapper.to_response at the HTTP boundary
- Removes IEEE 754 floating-point precision anti-pattern from template

Pagination in list_items:
- ListItemsQuery adds limit (default 50) and offset (default 0)
- IItemRepository.find_all(limit, offset) — port updated with pagination
- InMemoryItemRepository implements slice-based pagination
- ListItemsHandler passes pagination from query to repository
- IItemApplicationService.list_items(limit, offset)
- GET /api/v1/items now accepts ?limit=N&offset=M query params

mark_deleted() semantics fixed:
- Entity base class gains is_deleted: bool = field(init=False, default=False)
- Item.mark_deleted() sets self.is_deleted = True + emits event
- Soft-delete state now observable on the aggregate

DELETE idempotency:
- DeleteItemHandler: absent item returns Success(None) instead of Failure
- HTTP DELETE now correctly returns 204 for both existing and absent items
- Aligns with REST idempotency semantics

__all__ in public module interfaces:
- application/ports/item_application_service.py: __all__ = [...]
- application/dtos/auth_dtos.py: __all__ = [...]

Tests updated:
- All unit tests updated to use Decimal
- Added test_mark_deleted_sets_flag (domain)
- Added test_list_items_pagination (handlers)
- test_delete_nonexistent_returns_failure → test_delete_nonexistent_is_idempotent
- HTTP tests: added pagination test; DELETE-not-found now expects 204

All checks: 39 tests, pylint 10/10, mypy 70 files 0 errors, tach ✅, import-linter 3/3 ✅

+++ src/app/application/commands/item_commands.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 import uuid
 from dataclasses import dataclass, field
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
@@ -16,7 +17,7 @@ class CreateItemCommand:
     """Command to create a new item."""
 
     name: str
-    price: float
+    price: Decimal
     description: str = ""
     command_id: uuid.UUID = field(default_factory=uuid.uuid4)
 
@@ -27,7 +28,7 @@ class UpdateItemCommand:
 
     item_id: uuid.UUID
     name: str | None = None
-    price: float | None = None
+    price: Decimal | None = None
     description: str | None = None
     command_id: uuid.UUID = field(default_factory=uuid.uuid4)
